now supports multipart/form-data (as far as this project is concerned, atm) now using send() rather than write() so broken pipe doesn't terminate the process

+++ src/network.h
@@ -5,14 +5,14 @@
 struct client_state {
 	int socket;
 	char* buffer;
-	int size;
-	int allocated;
+	size_t size;
+	size_t allocated;
 	struct http_headers headers;
 	bool has_headers;
 };
 
 void initialize_network();
 void poll_network();
-void read_from_client(int socket, char** buffer, int* size, int* allocated, int max_size);
-void socket_write_all(int socket, const char* buffer, int size);
+void read_from_client(int socket, char** buffer, size_t* size, size_t* allocated, size_t max_size);
+void socket_write_all(int socket, const char* buffer, size_t size);
 void free_network();
